Renamed Ogre.js file
#8
