+++ src/Monsters/Races/Monsters/Ogres.js
