+++ src/Monsters/Tables/Level1.js
@@ -12,7 +12,7 @@ import {GiantBat} from "../Races/Monsters/GiantBat";
 import {GiantRat} from "../Races/Monsters/GiantRats";
 import {GiantSpider} from "../Races/Monsters/GiantSpiders";
 import {Minotaur} from "../Races/Monsters/Minotaurs";
-import {Ogre} from "../Races/Monsters/Ogre";
+import {Ogre} from "../Races/Monsters/Ogres";
 
 import Skaven from "../Races/Skaven";
 import {Clanrat} from "../Races/Skaven/Clanrat";
